hdrnet/requirements.txt
numpy
pyglib
scikit_image
python_gflags
python_magic
setproctitle
tensorflow-gpu==2.9.3
opencv-python
pytest